Refactor coordinate system to float and improve visualization controls
This commit refactors the coordinate system from int to float throughout
the codebase for improved precision in bee movement and environment
representation. It also adds interactive visualization controls and tunnel
entrance/exit rendering.

Key changes:
- Convert environment and tunnel dimensions from int to float
- Encapsulate Hive member variables with accessor methods
- Add tunnel collision detection (inTunnel method) with wall thickness
- Implement tunnel entrance visualization with configurable boundaries
- Add keyboard controls for simulation speed (+/- keys) and EMD display (E key)
- Display simulation speed in status text
- Add version string macro to polybeeConfig.h.in
- Fix heatmap cell size warnings to handle float environment dimensions
- Refactor bee drawing to use envToDisplay coordinate conversion consistently
- Improve status display with adjustable font sizes and EMD toggle

+++ include/Params.h
@@ -78,14 +78,14 @@ public:
     static int numIterations;
 
     // Environment configuration
-    static int envW;
-    static int envH;
+    static float envW;
+    static float envH;
 
     // Tunnel configuration
-    static int tunnelW;
-    static int tunnelH;
-    static int tunnelX; // top-left x position of tunnel in environment coordinates
-    static int tunnelY; // top-left y position of tunnel in environment coordinates
+    static float tunnelW;
+    static float tunnelH;
+    static float tunnelX; // top-left x position of tunnel in environment coordinates
+    static float tunnelY; // top-left y position of tunnel in environment coordinates
     static std::vector<TunnelEntranceSpec> tunnelEntranceSpecs;
 
     // Bee configuration
